Move the offline command to the top level
